@@ -10,32 +10,38 @@
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --no-progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js"
     },
     "devDependencies": {
-        "ajv": "^6.5.2",
-        "ajv-keywords": "^3.2.0",
-        "axios": "^0.18",
-        "bootstrap-sass": "^3.3.7",
-        "cross-env": "^5.1",
-        "dropzone": "^4.3.0",
-        "epubjs": "^0.4.2",
-        "featherlight": "^1.7.13",
-        "jqtree": "^1.4.7",
-        "jquery": "^3.3.1",
+        "@tailwindcss/forms": "^0.2.1",
+        "alpinejs": "^2.7.3",
+        "autoprefixer": "^10.1.0",
+        "axios": "^0.21.4",
+        "laravel-mix": "^6.0.6",
+        "lodash": "^4.17.19",
+        "postcss": "^8.2.1",
+        "postcss-import": "^12.0.1",
+        "resolve-url-loader": "^4.0.0",
+        "sass": "^1.37.0",
+        "sass-loader": "^12.1.0",
+        "tailwindcss": "^2.0.2"
+    },
+    "dependencies": {
+        "bootstrap-sass": "3.4.1",
+        "dropzone": "^5.9.2",
+        "featherlight": "^1.7.14",
+        "jqtree": "^1.6.1",
+        "jquery": "^3.6.0",
         "jquery-resizable": "^1.0.6",
-        "jquery-ui": "^1.12.1",
-        "laravel-mix": "^2.1.11",
-        "lodash": "^4.17.10",
-        "popper.js": "^1.12",
+        "jquery-ui": "^1.13.0",
+        "node-sass": "^6.0.1",
+        "nth-check": "^2.0.1",
+        "object-path": "^0.11.8",
+        "path-parse": "^1.0.7",
         "selectize": "^0.12.6",
         "simplemde": "^1.11.2",
         "snackbar": "^1.1.0",
-        "v-calendar": "^0.9.7",
-        "vue": "^2.5.16",
-        "vue-loader": "^14.2.3"
-    },
-    "dependencies": {
-        "fs": "0.0.1-security",
-        "grunt": "^1.0.3",
-        "pegjs": "^0.10.0",
-        "reveal": "0.0.4"
+        "url-parse": "^1.5.4",
+        "v-calendar": "^2.3.2",
+        "vue": "^2.6.14",
+        "vue-loader": "^15.9.7",
+        "vue-template-compiler": "^2.6.14"
     }
 }
