@@ -8,8 +8,11 @@
 require('./bootstrap');
 require('selectize');
 
-var SimpleMDE = require('simplemde/dist/simplemde.min');
-window.Dropzone = require('dropzone');
+//var SimpleMDE = require('simplemde/dist/simplemde.min');
+import * as SimpleMDE from 'simplemde/dist/simplemde.min';
+
+//window.Dropzone = require('dropzone');
+import * as Dropzone from 'dropzone';
 
 require('jquery-ui/ui/widgets/autocomplete');
 
@@ -21,42 +24,38 @@ require('jqtree');
 
 window.snackbar = require('snackbar');
 
-Vue.component('shares', require('./components/Shares.vue'));
+import Vue from 'vue';
+window.Vue = Vue;
 
-Vue.component(
-    'passport-clients',
-    require('./components/passport/Clients.vue')
-);
+import Shares from './components/Shares.vue';
+Vue.component('shares', Shares);
 
-Vue.component(
-    'passport-authorized-clients',
-    require('./components/passport/AuthorizedClients.vue')
-);
+import Clients from './components/passport/Clients.vue';
+Vue.component('passport-clients', Clients);
 
-Vue.component(
-    'passport-personal-access-tokens',
-    require('./components/passport/PersonalAccessTokens.vue')
-);
+import AuthorizedClients from './components/passport/AuthorizedClients.vue';
+Vue.component('passport-authorized-clients', AuthorizedClients);
 
-/*
+import PersonalAccessTokens from './components/passport/PersonalAccessTokens.vue';
+Vue.component('passport-personal-access-tokens', PersonalAccessTokens);
+
 import VCalendar from 'v-calendar';
 Vue.use(VCalendar, {
     firstDayOfWeek: 2
 });
 
-Vue.component(
-    'mini-calendar',
-    require('./components/Calendar.vue')
-);
-*/
+import Calendar from './components/Calendar.vue';
+Vue.component('mini-calendar', Calendar);
 
 window.markdownEditor = function () {
+    var uuid = $('meta[name="uuid"]').attr('content')
+        + '-' + $('meta[name="updated"]').attr('content').replace(/[-: ]/g, '');
     var simplemde = new SimpleMDE({
         element: $("#body-input")[0],
         autofocus: true,
         autosave: {
             enabled: true,
-            uniqueId: $('meta[name="uuid"]').attr('content')
+            uniqueId: uuid
         },
         spellChecker: false,
         toolbar: [
