@@ -9,7 +9,7 @@ class UpdateTagsTable extends Migration {
 	{
 
 		Schema::table('tagging_tags', function ($table) {
-			$table->integer('tag_group_id')->unsigned()->nullable()->after('id');
+			$table->bigInteger('tag_group_id')->unsigned()->nullable()->after('id');
 			$table->foreign('tag_group_id')->references('id')->on('tagging_tag_groups');
 		});
 
