@@ -14,11 +14,11 @@ class CreateItemsTable extends Migration
     public function up()
     {
         Schema::create('items', function (Blueprint $table) {
-            $table->increments('id');
+            $table->id();
             $table->string('title');
-            $table->integer('concept_id')->unsigned();
+            $table->bigInteger('concept_id')->unsigned();
             $table->foreign('concept_id')->references('id')->on('concepts');
-            $table->integer('owner_id')->unsigned();
+            $table->bigInteger('owner_id')->unsigned();
             $table->foreign('owner_id')->references('id')->on('users');
             $table->boolean('is_done')->default(false)->index();
             $table->date('due_at')->nullable()->index();
