@@ -1,4 +1,4 @@
-@extends('layouts.app')
+@extends('knowfox::layouts.app')
 
 @section('content')
 
@@ -11,7 +11,7 @@
             </ol>
 
             <h1>{{$page_title}} <small>{{$sub_title}}</small></h1>
-            @include('partials.messages')
+            @include('knowfox::partials.messages')
 
         </section>
 
