@@ -1,4 +1,4 @@
-@extends('concept.show')
+@extends('knowfox::concept.show')
 
 @section('kids-header')
     <thead>
@@ -20,8 +20,8 @@
                     <img src="/{{$book->id}}/{{$book->config->image}}?style=thumbnail">
                 @endif
             </td>
-            <td>{{$book->config->author or ''}}</td>
-            <td><a href="{{route('concept.show', ['concept' => $book])}}">
+            <td>{{$book->config->author ?? ''}}</td>
+            <td><div><a href="{{route('concept.show', ['concept' => $book])}}">
                     {{$book->title}}
                 </a>
                 @if ($book->is_flagged)
@@ -32,8 +32,10 @@
                     {!! $sep !!}<a href="{{route('concept.index', ['tag' => $tag->slug])}}" class="label label-default">{{$tag->name}}</a>
                     <?php $sep = ''; ?>
                 @endforeach
+                </div>
+                <div>{{ $book->summary }}</div>
             </td>
-            <td>{{$book->config->year or ''}}</td>
+            <td>{{$book->config->year ?? ''}}</td>
         </tr>
     @endforeach
     </tbody>
