@@ -1,7 +1,7 @@
-@extends('concept.show')
+@extends('knowfox::concept.show')
 
 @section('top-sidebar')
-    <!-- mini-calendar></mini-calendar -->
+    <mini-calendar></mini-calendar>
 @endsection
 
 @section('siblings')
