@@ -1,4 +1,4 @@
-@extends('layouts.app')
+@extends('knowfox::layouts.app')
 
 @section('content')
 
@@ -12,7 +12,7 @@
 
             <h1>{{$page_title}}</small></h1>
 
-            @include('partials.messages')
+            @include('knowfox::partials.messages')
 
         </section>
 
