@@ -3,7 +3,7 @@
         <input type="hidden" name="concept_id" value="{{ $concept->id }}">
     @endif
     <div class="input-group">
-        <input id="search-input" type="search" name="q" class="form-control" value="{{ $search_term or '' }}" placeholder="Search {{ $concept->title or '' }}">
+        <input id="search-input" type="search" name="q" class="form-control" value="{{ isset($search_term) ? $search_term : '' }}" placeholder="Search {{ isset($concept) ? $concept->title : '' }}">
         <div class="input-group-btn">
             <button class="btn btn-default" type="button"><i class="glyphicon glyphicon-search"></i></button>
             @if (!empty($concept))
