@@ -16,7 +16,7 @@ Route::get('/', function () {
     if (Auth::check()) {
         return redirect()->route('home');
     }
-    return view('welcome');
+    return view('knowfox::welcome');
 });
 
 Route::get('cancel/{what}/{email}', [
@@ -24,7 +24,7 @@ Route::get('cancel/{what}/{email}', [
     'uses' => 'UserController@cancel',
 ]);
 
-Auth::routes();
+//Auth::routes();
 
 Route::get('auth/email-authenticate/{token}/{cid?}', [
     'as' => 'auth.email-authenticate',
